+++ CHANGES.md
@@ -2,6 +2,7 @@
 * Added SignConfiguration POST API
 * Added SignConfiguration PUT API
 * Added SignConfiguration GET API
+* Added SignConfiguration DELETE API
 
 1.0.0
 * Created a nuget package